Redundant typecast, member reorganization etc.

+++ src/org/ankhzet/ergo/ui/xgui/XPathFilePicker.java
@@ -133,7 +133,7 @@ public class XPathFilePicker extends CommonControl {
   }
 
   int rowsInView() {
-    return (int) ((h - ITEM_HEIGHT) / ITEM_HEIGHT);
+    return ((h - ITEM_HEIGHT) / ITEM_HEIGHT);
   }
 
   float columnWidth() {
